// FE/src/types/User.ts
export interface User {
  _id?: string;
    clerkId: string;
    firstName: string;
    lastName: string;
    imageUrl: string;
    email: string;
    Address:string;
    role: string;
    phone: string;
    gender: string;
    isActive: string;
    paymentInfo: string;
    orders:string;
    isBanned: boolean;
    
  }